data/ref_projects/cepri-dev-new/requirements.txt
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.7.1
et-xmlfile==1.1.0
fire==0.6.0
gevent==24.2.1
greenlet==3.0.3
idna==3.10
IPy==1.1
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
paramiko==3.4.0
prettytable==3.11.0
prompt-toolkit==2.0.10
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
whl/3.3.0/RenixLibrary-3.3.0.102344-py3-none-any.whl
requests==2.32.3
robotframework==4.1.3
robotframework-debuglibrary==2.2.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
tabulate==0.9.0
termcolor==2.4.0
whl/3.3.0/TesterLibrary-3.3.0.102344-py3-none-any.whl
tzdata==2024.1
untangle==1.2.1
urllib3==2.2.3
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.0.3
whl/3.3.0/CustomLibrary-3.3.0.102431-py3-none-any.whl
